+++ src/components/modals/ChatSessionModal.tsx
@@ -12,7 +12,7 @@ import Webchat, { renderActivity } from '../Webchat'
 import { returntypeof } from 'react-redux-typescript'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { EditDialogType } from '.'
+import { EditDialogType } from '../../types/const'
 import { State } from '../../types'
 import { AppBase } from '@conversationlearner/models'
 import { FM } from '../../react-intl-messages'
