+++ src/components/modals/EditDialogModal.tsx
@@ -21,9 +21,9 @@ import Webchat, { renderActivity } from '../Webchat'
 import { ImportedAction } from '../../types/models'
 import { formatMessageId, equal, deepCopy } from '../../Utils/util'
 import { State } from '../../types'
-import { EditDialogAdmin, EditDialogType, EditState } from '.'
+import { EditDialogAdmin } from '.'
 import { Activity } from 'botframework-directlinejs'
-import { SelectionType } from '../../types/const'
+import { EditDialogType, EditState, SelectionType } from '../../types/const'
 import { returntypeof } from 'react-redux-typescript'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
