+++ src/components/modals/EntityExtractor.tsx
@@ -13,7 +13,7 @@ import actions from '../../actions'
 import HelpIcon from '../HelpIcon'
 import EntityCreatorEditor from './EntityCreatorEditor'
 import { injectIntl, InjectedIntlProps } from 'react-intl'
-import { EditDialogType } from '.'
+import { EditDialogType } from '../../types/const'
 import { FM } from '../../react-intl-messages'
 import { returntypeof } from 'react-redux-typescript'
 import { bindActionCreators } from 'redux'
