+++ src/components/modals/TeachSessionAdmin.tsx
@@ -16,7 +16,7 @@ import { returntypeof } from 'react-redux-typescript'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { State } from '../../types'
-import { EditDialogType } from '../../components/modals'
+import { EditDialogType } from '../../types/const'
 import { FM } from '../../react-intl-messages'
 import { TeachSessionState } from '../../types/StateTypes'
 import { injectIntl, InjectedIntlProps } from 'react-intl'
