+++ src/routes/Apps/App/LogDialogs.tsx
@@ -18,8 +18,8 @@ import { returntypeof } from 'react-redux-typescript'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { State } from '../../../types'
-import { SelectionType } from '../../../types/const'
-import { ChatSessionModal, EditDialogModal, TeachSessionModal, EditDialogType, EditState, MergeModal, ConfirmCancelModal } from '../../../components/modals'
+import { EditDialogType, EditState, SelectionType } from '../../../types/const'
+import { ChatSessionModal, EditDialogModal, TeachSessionModal, MergeModal, ConfirmCancelModal } from '../../../components/modals'
 import LogConversionConflictModal, { ConflictPair } from '../../../components/modals/LogConversionConflictModal'
 import { injectIntl, InjectedIntl, InjectedIntlProps } from 'react-intl'
 import { FM } from '../../../react-intl-messages'
