+++ src/routes/Apps/App/TrainDialogs.tsx
@@ -25,8 +25,8 @@ import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { returntypeof } from 'react-redux-typescript'
 import { State, ErrorType } from '../../../types'
-import { SelectionType, FeatureStrings } from '../../../types/const'
-import { TeachSessionModal, EditDialogModal, EditDialogType, EditState, MergeModal } from '../../../components/modals'
+import { EditDialogType, EditState, SelectionType, FeatureStrings } from '../../../types/const'
+import { TeachSessionModal, EditDialogModal, MergeModal } from '../../../components/modals'
 import { injectIntl, InjectedIntl, InjectedIntlProps } from 'react-intl'
 import { FM } from '../../../react-intl-messages'
 import { Activity } from 'botframework-directlinejs'
